+++ Cargo.toml
@@ -53,6 +53,7 @@ insta = { version = "1.29.0", features = ["yaml"] }
 integer-encoding = "3.0.4"
 jsonrpsee = "0.18.1"
 jsonschema = "0.17.0"
+lazy_static = "1.4.0"
 libmdbx = ">=0.1.5,<0.1.8"
 metrics = "0.21.0"
 mockall = "0.11.2"
