+++ crates/papyrus_gateway/src/broadcasted_transaction.rs
@@ -5,6 +5,9 @@ use starknet_writer_client::objects::transaction::{
     DeclareV1Transaction, DeployAccountTransaction, InvokeTransaction,
 };
 
+// DeclareV2 is not from the client because the broadcasted transaction of declare v2 has
+// slight alterations from the client declare v2. We define our own BroadcastedDeclareV2Transaction
+// further below.
 pub type BroadcastedDeclareV1Transaction = DeclareV1Transaction;
 pub type BroadcastedDeployAccountTransaction = DeployAccountTransaction;
 pub type BroadcastedInvokeTransaction = InvokeTransaction;
@@ -15,8 +18,6 @@ use crate::v0_3_0::state::ContractClass;
 #[serde(tag = "type")]
 pub enum BroadcastedTransaction {
     #[serde(rename = "DECLARE")]
-    // Declare is not from the client because the broadcasted transaction of declare has slight
-    // alterations from the client declare.
     Declare(BroadcastedDeclareTransaction),
     #[serde(rename = "DEPLOY_ACCOUNT")]
     DeployAccount(BroadcastedDeployAccountTransaction),
@@ -28,8 +29,6 @@ pub enum BroadcastedTransaction {
 #[serde(untagged)]
 pub enum BroadcastedDeclareTransaction {
     DeclareV1(BroadcastedDeclareV1Transaction),
-    // DeclareV2 is not from the client because the broadcasted transaction of declare v2 has
-    // slight alterations from the client declare v2.
     DeclareV2(BroadcastedDeclareV2Transaction),
 }
