+++ crates/starknet_writer_client/src/objects/mod.rs
@@ -1,3 +1,5 @@
+#[cfg(any(feature = "testing", test))]
+pub mod test_utils;
 pub mod transaction;
 #[cfg(test)]
 mod transaction_test;
